Rename transcribe endpoint to transcription
