+++ main.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from app.config import settings
-from app.api.transcribe import router as api_router
+from app.api.transcription import router as api_router
 
 app = FastAPI(title="Voice Transcription API")
